Move cli.py to app dir
